+++ src/install.js
@@ -203,5 +203,9 @@ module.exports = async (runtime, cb) => {
     (npm run build:production || npm run web:build:production) && (npm run deploy || npm run web:deploy)
   `, { 'encoding': 'utf-8', 'mode': '755' })
 
+  if (runtime.lanyonUpdateGemLockfile === true) {
+    utils.fsCopySync(`${runtime.cacheDir}/Gemfile.lock`, `${runtime.lanyonDir}/Gemfile.lock`)
+  }
+
   cb(null)
 }