+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Released: TBA.
 
 - [ ] Add a real command line parser for `lib/cli.js` like minimist
 - [ ] Consider not testing each install setup, but just verifying at least one works
+- [x] Distribute the `Gemfile.lock` as well
 - [x] Switch from ava -> jest
 - [x] Fix bug: `the input device is not a TTY`
 - [x] No longer auto install ruby things upon `install`, require an explicit call to `cli.js install` instead